exclude game_engine module
