+++ brain_games/games/brain_prime.py
@@ -1,6 +1,7 @@
 """Prime number game."""
 
-from brain_games.common import generate_random_int, run_game
+from brain_games.common import generate_random_int
+from brain_games.game_engine import run_game
 
 
 def brain_prime() -> None:
